requirements.txt
streamlit
pandas
openpyxl
python-barcode
pillow